+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/payment/PaymentDto.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.dto.payment;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -15,8 +16,8 @@ import java.util.UUID;
 @Builder
 public record PaymentDto (
         UUID paymentId,
-        Double totalPayment,
-        Double deliveryTotal,
-        Double feeTotal
+        BigDecimal totalPayment,
+        BigDecimal deliveryTotal,
+        BigDecimal feeTotal
 ) {
 }
